refactor[netOperatingProfit 3] Extract both parts in metrics and component
- The computation of the data is part of the metrics (so it can be available to comparison later)
- The rest is display so computed by the component
