+++ src/utils/data/metrics/index.js
@@ -1,5 +1,6 @@
 import { buildBenefitCostRatioData } from "./eco/benefitCostRatio.js";
 import { buildGlobalEmploymentData } from "./eco/employment.js";
+import { buildNetOperatingProfitPerActor } from "./eco/netOperatingProfitPerActor.js";
 
 export function computeMetrics(studyData) {
   return {
@@ -12,6 +13,7 @@ function buildEcoMetrics(studyData) {
 
   return {
     benefitCostRatio: buildBenefitCostRatioData(studyData.ecoData),
-    employment: buildGlobalEmploymentData(studyData.ecoData)
+    employment: buildGlobalEmploymentData(studyData.ecoData),
+    netOperatingProfitPerActor: buildNetOperatingProfitPerActor(studyData.ecoData)
   };
 }
